requirements.txt
aiogram==2.19
APScheduler==3.9.1
google_api_python_client==2.39.0
google_auth_oauthlib==0.4.5
httplib2==0.20.4
pandas==1.3.4
phonenumbers==8.12.44
protobuf==3.19.4
python-dotenv==0.19.2
requests==2.26.0
SQLAlchemy==1.4.31
Telethon==1.24.0
tqdm==4.62.3
